requirements.txt
asyncio
httpx
pydantic
python-dotenv
regex